Use a single directory for clang's profile data output.
For some reason in CPython, but not PyPy, using different directories
with multiple extensions would only output to the directory of the
first extension loaded by profiling code, so the build_ext_profile_use
command would think that profile data was missing. I experimented with
a bunch of different compile flag overrides to see if one of them had
any effect, but without any results. The least hacky solution seemed
to be this anyway.